// src/api/model/apiProcessDocument.ts
/*
 * Copyright(c) 2009 - 2019 Abelium d.o.o.
 * Kajuhova 90, 1000 Ljubljana, Slovenia
 * All rights reserved
 * Copyright(c) 1995 - 2018 T-Systems Multimedia Solutions GmbH
 * Riesaer Str. 5, 01129 Dresden
 * All rights reserved.
 *
 * INATrace Services API
 * Abelium INATrace Services API swagger documentation
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the openapi-typescript-angular-generator.
 * https://github.com/alenabelium/openapi-typescript-angular-generator
 * Do not edit the class manually.
 */


import { ApiDocument } from './apiDocument';



export interface ApiProcessDocument { 
    /**
     * description of this document
     */
    description?: string;
    document?: ApiDocument;
}

/**
 * Namespace for property- and property-value-enumerations of ApiProcessDocument.
 */
export namespace ApiProcessDocument {
    /**
     * All properties of ApiProcessDocument.
     */
    export enum Properties {
        /**
         * description of this document
         */
        description = 'description',
        document = 'document'
    }


    export function formMetadata() {
        return  {
            metadata: formMetadata,
            classname: 'ApiProcessDocument',
            vars: [
                {
                    isReadOnly: false,
                    isEnum: false,
                    required: false,
                    name: 'description',
                    classname: 'ApiProcessDocument',
                    dataType: 'string',
                    isPrimitiveType: true,
                    isListContainer: false,
                    complexType: ''
                },
                {
                    metadata: ApiDocument.formMetadata,
                    isReadOnly: false,
                    isEnum: false,
                    required: false,
                    name: 'document',
                    classname: 'ApiProcessDocument',
                    dataType: 'ApiDocument',
                    isPrimitiveType: false,
                    isListContainer: false,
                    complexType: 'ApiDocument'
                },
            ],
            validators: {
                description: [
                ],
                document: [
                ],
            }
        }
    }

  // export const ApiProcessDocumentValidationScheme = {
  //     validators: [],
  //     fields: {
  //               description: {
  //                   validators: []
  //               },
  //               document: {
  //                   validators: []
  //               },
  //     }
  // } as SimpleValidationScheme<ApiProcessDocument>;


}